refactor(vscode): split opencode.ts into focused modules
Split the monolithic OpenCode manager into types, auth, cli-discovery,
env, readiness, spawn, and manager modules. Keep opencode.ts as a thin
barrel for backward-compatible imports.

Behavior improvements:
- Probe external configured apiUrl health before marking connected
- CLI missing dialog offers Open Settings, Retry, and More Info actions
- onStatusChange callback now includes optional cliAvailable meta
- waitForReady uses a lazy singleton OutputChannel

Add focused unit tests for cli-discovery, readiness, and external URL
startup.
