+++ packages/vscode/src/opencode/env.ts
@@ -1,0 +1,132 @@
+import * as path from 'path';
+import { spawnSync } from 'child_process';
+
+let cachedLoginShellEnvSnapshot: Record<string, string> | null | undefined;
+
+function parseNullSeparatedEnvSnapshot(raw: string): Record<string, string> | null {
+  if (typeof raw !== 'string' || raw.length === 0) {
+    return null;
+  }
+
+  const result: Record<string, string> = {};
+  const entries = raw.split('\0');
+  for (const entry of entries) {
+    if (!entry) {
+      continue;
+    }
+    const idx = entry.indexOf('=');
+    if (idx <= 0) {
+      continue;
+    }
+    const key = entry.slice(0, idx);
+    const value = entry.slice(idx + 1);
+    result[key] = value;
+  }
+
+  return Object.keys(result).length > 0 ? result : null;
+}
+
+function getWindowsShellEnvSnapshot(): Record<string, string> | null {
+  const parseResult = (stdout: string | null | undefined) => parseNullSeparatedEnvSnapshot(typeof stdout === 'string' ? stdout : '');
+
+  const psScript =
+    "Get-ChildItem Env: | ForEach-Object { [Console]::Out.Write($_.Name); [Console]::Out.Write('='); [Console]::Out.Write($_.Value); [Console]::Out.Write([char]0) }";
+
+  const powershellCandidates = [
+    'pwsh.exe',
+    'powershell.exe',
+    path.join(process.env.SystemRoot || 'C:\\Windows', 'System32', 'WindowsPowerShell', 'v1.0', 'powershell.exe'),
+  ];
+
+  for (const shellPath of powershellCandidates) {
+    try {
+      const result = spawnSync(shellPath, ['-NoLogo', '-Command', psScript], {
+        encoding: 'utf8',
+        stdio: ['ignore', 'pipe', 'pipe'],
+        maxBuffer: 10 * 1024 * 1024,
+        windowsHide: true,
+      });
+      if (result.status !== 0) {
+        continue;
+      }
+      const parsed = parseResult(result.stdout);
+      if (parsed) {
+        return parsed;
+      }
+    } catch {
+      continue;
+    }
+  }
+
+  const comspec = process.env.ComSpec || 'cmd.exe';
+  try {
+    const result = spawnSync(comspec, ['/d', '/s', '/c', 'set'], {
+      encoding: 'utf8',
+      stdio: ['ignore', 'pipe', 'pipe'],
+      maxBuffer: 10 * 1024 * 1024,
+      windowsHide: true,
+    });
+    if (result.status === 0 && typeof result.stdout === 'string' && result.stdout.length > 0) {
+      return parseNullSeparatedEnvSnapshot(result.stdout.replace(/\r?\n/g, '\0'));
+    }
+  } catch {
+    return null;
+  }
+
+  return null;
+}
+
+function getLoginShellEnvSnapshot(): Record<string, string> | null {
+  if (cachedLoginShellEnvSnapshot !== undefined) {
+    return cachedLoginShellEnvSnapshot;
+  }
+
+  // Avoid interactive POSIX login shells in the extension host.
+  if (process.platform !== 'win32') {
+    cachedLoginShellEnvSnapshot = null;
+    return null;
+  }
+
+  const windowsSnapshot = getWindowsShellEnvSnapshot();
+  cachedLoginShellEnvSnapshot = windowsSnapshot;
+  return windowsSnapshot;
+}
+
+export function mergePathValues(preferred: string, fallback: string): string {
+  const merged = new Set<string>();
+  const addSegments = (value: string) => {
+    if (typeof value !== 'string' || !value) {
+      return;
+    }
+    for (const segment of value.split(path.delimiter)) {
+      if (segment) {
+        merged.add(segment);
+      }
+    }
+  };
+
+  addSegments(preferred);
+  addSegments(fallback);
+  return Array.from(merged).join(path.delimiter);
+}
+
+export function applyLoginShellEnvSnapshot() {
+  const snapshot = getLoginShellEnvSnapshot();
+  if (!snapshot) {
+    return;
+  }
+
+  const skipKeys = new Set(['PWD', 'OLDPWD', 'SHLVL', '_']);
+  for (const [key, value] of Object.entries(snapshot)) {
+    if (skipKeys.has(key)) {
+      continue;
+    }
+    const existing = process.env[key];
+    if (typeof existing === 'string' && existing.length > 0) {
+      continue;
+    }
+    process.env[key] = value;
+  }
+
+  process.env.PATH = mergePathValues(snapshot.PATH || '', process.env.PATH || '');
+}
